context_processor/requirements.txt
paho-mqtt
jsonschema
pandas
json-logging